fix: Fixed migration issue for react router v6; renamed Dashboard.tsx to index for clearer code, fixed login on valid user logic
